Deduplicate invoking flow-schema-inference
This was all essentially the same in both Kinesis and filesource.
There's no real need for that.
