+++ source-kinesis/discovery.go
@@ -4,23 +4,14 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"io"
-	"os/exec"
 	"sync"
 	"time"
 
 	"github.com/aws/aws-sdk-go/service/kinesis"
+	"github.com/estuary/connectors/schema_inference"
 	"github.com/estuary/flow/go/protocols/airbyte"
 	log "github.com/sirupsen/logrus"
-	"golang.org/x/sync/errgroup"
 )
-
-// How long should we allow reading from each stream to run. Note: empty streams
-// will wait for documents until this timeout triggers.
-const DISCOVER_READ_TIMEOUT = time.Second * 10
-
-// How long should we allow schema inference to run.
-const DISCOVER_INFER_TIMEOUT = time.Second * 5
 
 // How many documents to read from each stream. A higher number gives us more
 // opportunity to spot variation in the document stream at the expense of having
@@ -28,7 +19,7 @@ const DISCOVER_INFER_TIMEOUT = time.Second * 5
 const DISCOVER_MAX_DOCS_PER_STREAM = 100
 
 // Provides a default schema to use when we encounter an issue with schema inference.
-var FALLBACK_SCHEMA = json.RawMessage(`{"type":"object"}`)
+var DISCOVER_FALLBACK_SCHEMA = json.RawMessage(`{"type":"object"}`)
 
 func discoverStreams(ctx context.Context, client *kinesis.Kinesis, streamNames []string) []airbyte.Stream {
 	var waitGroup = new(sync.WaitGroup)
@@ -47,12 +38,12 @@ func discoverStreams(ctx context.Context, client *kinesis.Kinesis, streamNames [
 				// looks wrong to the user, they can investigate, but ultimately
 				// not every stream will contain json documents.
 				log.WithField("stream", name).Error(err.Error())
-				schema = &FALLBACK_SCHEMA
+				schema = DISCOVER_FALLBACK_SCHEMA
 			}
 
 			streams[i] = airbyte.Stream{
 				Name:                name,
-				JSONSchema:          *schema,
+				JSONSchema:          schema,
 				SupportedSyncModes:  []airbyte.SyncMode{airbyte.SyncModeIncremental},
 				SourceDefinedCursor: true,
 			}
@@ -64,25 +55,10 @@ func discoverStreams(ctx context.Context, client *kinesis.Kinesis, streamNames [
 	return streams
 }
 
-func discoverSchema(ctx context.Context, client *kinesis.Kinesis, streamName string) (*json.RawMessage, error) {
-	var logEntry = log.WithField("stream", streamName)
-	var peekCtx, cancelPeek = context.WithTimeout(ctx, DISCOVER_READ_TIMEOUT)
-	defer cancelPeek()
-
-	var docCh, docCount, err = peekAtStream(peekCtx, client, streamName, DISCOVER_MAX_DOCS_PER_STREAM)
-	if err != nil {
-		return nil, fmt.Errorf("schema discovery for stream `%s` failed: %w", streamName, err)
-	} else if docCount == 0 {
-		return nil, fmt.Errorf("no documents discovered for stream: %v", streamName)
-	}
-
-	logEntry.WithField("count", docCount).Info("Got documents for stream")
-
-	var inferCtx, cancelInfer = context.WithTimeout(ctx, DISCOVER_INFER_TIMEOUT)
-	defer cancelInfer()
-
-	var schema *json.RawMessage
-	schema, err = runInference(inferCtx, logEntry, docCh)
+func discoverSchema(ctx context.Context, client *kinesis.Kinesis, streamName string) (json.RawMessage, error) {
+	schema, err := schema_inference.Run(ctx, streamName, func(ctx context.Context, logEntry *log.Entry) (<-chan json.RawMessage, uint, error) {
+		return peekAtStream(ctx, client, streamName, DISCOVER_MAX_DOCS_PER_STREAM)
+	})
 	if err != nil {
 		return nil, fmt.Errorf("failed to infer schema for stream `%s`: %w", streamName, err)
 	}
@@ -90,71 +66,7 @@ func discoverSchema(ctx context.Context, client *kinesis.Kinesis, streamName str
 	return schema, nil
 }
 
-func runInference(ctx context.Context, logEntry *log.Entry, docsCh chan json.RawMessage) (*json.RawMessage, error) {
-	var (
-		err      error
-		errGroup *errgroup.Group
-		schema   = new(json.RawMessage)
-		cmd      *exec.Cmd
-		stdin    io.WriteCloser
-		stdout   io.ReadCloser
-	)
-
-	errGroup, ctx = errgroup.WithContext(ctx)
-	cmd = exec.CommandContext(ctx, "flow-schema-inference", "analyze")
-
-	stdin, err = cmd.StdinPipe()
-	if err != nil {
-		return nil, fmt.Errorf("failed to open stdin: %w", err)
-	}
-	stdout, err = cmd.StdoutPipe()
-	if err != nil {
-		return nil, fmt.Errorf("failed to open stdout: %w", err)
-	}
-
-	errGroup.Go(func() error {
-		defer stdin.Close()
-		encoder := json.NewEncoder(stdin)
-
-		for doc := range docsCh {
-			if err := encoder.Encode(doc); err != nil {
-				return fmt.Errorf("failed to encode document: %w", err)
-			}
-		}
-		logEntry.Info("runInference: Done sending documents")
-
-		return nil
-	})
-
-	errGroup.Go(func() error {
-		defer stdout.Close()
-		decoder := json.NewDecoder(stdout)
-
-		if err := decoder.Decode(schema); err != nil {
-			return fmt.Errorf("failed to decode schema, %w", err)
-		}
-		logEntry.Debug("runInference: Done reading schema")
-
-		return nil
-	})
-
-	logEntry.Info("runInference: launching")
-	err = cmd.Start()
-	if err != nil {
-		return nil, fmt.Errorf("failed to run flow-schema-inference: %w", err)
-	}
-
-	errGroup.Go(cmd.Wait)
-
-	err = errGroup.Wait()
-	if err != nil {
-		return nil, err
-	}
-
-	return schema, nil
-}
-
-func peekAtStream(ctx context.Context, client *kinesis.Kinesis, streamName string, peekAtMost uint) (chan json.RawMessage, uint, error) {
+func peekAtStream(ctx context.Context, client *kinesis.Kinesis, streamName string, peekAtMost uint) (<-chan json.RawMessage, uint, error) {
 	var (
 		err         error
 		cancel      context.CancelFunc
